requirements.txt
streamlit
pypdf2
pdfplumber
pytesseract
pillow
openai
anthropic
pyyaml